requirements.txt
wordcloud==1.9.3
konlpy==0.6.0
selenium==4.19.0
schedule==1.2.1
textrankr==1.1
fastapi==0.103.0
openai-whisper==20231117
PyAudio==0.2.14
pandas
scipy
python-telegram-bot
python-dotenv
uvicorn